[#250] refactor: link 패키지 구조 변경
