+++ linkmind/src/main/java/com/app/toaster/toast/controller/response/WeekLinkDto.java
@@ -1,6 +1,6 @@
 package com.app.toaster.toast.controller.response;
 
-import com.app.toaster.domain.Link;
+import com.app.toaster.link.domain.Link;
 
 public record WeekLinkDto(Long linkId, String linkTitle, String linkImg, String linkUrl) {
     public static WeekLinkDto of(Link link){
